requirements.txt
praw
markdown
requests
python-dotenv
feedparser
sgmllib3k